refactor(http): split transport compatibility paths
