+++ pkg/services/recordings/transports/http/handlers_events.go
@@ -38,71 +38,122 @@ func (a *Adapter) GetEventsBySessionId(
 		Params:             params,
 		StreamGenerationID: r.Header.Get(SessionEventStreamGenerationHeader),
 	}
-	if isDurableHistorySession(input.SessionID) && a.hasLegacyHistory() {
-		if requestsJSONEventRecoveryProbe(r) {
-			a.probeLegacyEventRecovery(w, r, input.SessionID, input.Params)
-			return
-		}
-		stream, err := a.legacyEvents(r.Context(), input.SessionID, input.Params)
-		if shouldEndOnRequestContext(r.Context(), err) {
-			return
-		}
-		if err != nil {
-			if a.writeLegacyError(w, err, "failed to subscribe to factory events") {
-				return
-			}
-		}
-		a.streamLegacyFactoryEvents(w, r, stream, input.SessionID)
+	if a.handleLegacyDurableEvents(w, r, input) {
 		return
 	}
-	if !isDurableHistorySession(input.SessionID) && a.hasLegacyLiveEvents() {
-		if requestsJSONEventRecoveryProbe(r) {
-			a.probeLegacyLiveEventRecovery(w, r, input.SessionID, input.Params)
-			return
-		}
-		stream, err := a.legacyLive(r.Context(), input.SessionID, input.Params)
-		if shouldEndOnRequestContext(r.Context(), err) {
-			return
-		}
-		if err != nil {
-			a.writeLegacyError(w, err, "failed to subscribe to factory events")
-			return
-		}
-		a.streamLegacyFactoryEvents(w, r, stream, input.SessionID)
+	if a.handleLegacyLiveEvents(w, r, input) {
 		return
+	}
+	if a.handleEventRecoveryProbe(w, r, input) {
+		return
+	}
+	if a.handleHistoricalEvents(w, r, input) {
+		return
+	}
+	a.handleLiveEvents(w, r, input)
+}
+
+func (a *Adapter) handleLegacyDurableEvents(
+	w http.ResponseWriter,
+	r *http.Request,
+	input EventSubscribeInput,
+) bool {
+	if !isDurableHistorySession(input.SessionID) || !a.hasLegacyHistory() {
+		return false
 	}
 	if requestsJSONEventRecoveryProbe(r) {
-		if isDurableHistorySession(input.SessionID) {
-			a.probeHistoricalEventStreamRecovery(w, r, input)
-			return
-		}
-		a.probeEventStreamRecovery(w, input)
-		return
+		a.probeLegacyEventRecovery(w, r, input.SessionID, input.Params)
+		return true
+	}
+	stream, err := a.legacyEvents(r.Context(), input.SessionID, input.Params)
+	if shouldEndOnRequestContext(r.Context(), err) {
+		return true
+	}
+	if err != nil {
+		a.writeLegacyError(w, err, "failed to subscribe to factory events")
+		return true
+	}
+	a.streamLegacyFactoryEvents(w, r, stream, input.SessionID)
+	return true
+}
+
+func (a *Adapter) handleLegacyLiveEvents(
+	w http.ResponseWriter,
+	r *http.Request,
+	input EventSubscribeInput,
+) bool {
+	if isDurableHistorySession(input.SessionID) || !a.hasLegacyLiveEvents() {
+		return false
+	}
+	if requestsJSONEventRecoveryProbe(r) {
+		a.probeLegacyLiveEventRecovery(w, r, input.SessionID, input.Params)
+		return true
+	}
+	stream, err := a.legacyLive(r.Context(), input.SessionID, input.Params)
+	if shouldEndOnRequestContext(r.Context(), err) {
+		return true
+	}
+	if err != nil {
+		a.writeLegacyError(w, err, "failed to subscribe to factory events")
+		return true
+	}
+	a.streamLegacyFactoryEvents(w, r, stream, input.SessionID)
+	return true
+}
+
+func (a *Adapter) handleEventRecoveryProbe(
+	w http.ResponseWriter,
+	r *http.Request,
+	input EventSubscribeInput,
+) bool {
+	if !requestsJSONEventRecoveryProbe(r) {
+		return false
 	}
 	if isDurableHistorySession(input.SessionID) {
-		result, err := a.historicalRecording(r.Context(), input.SessionID)
-		if shouldEndOnRequestContext(r.Context(), err) {
-			return
-		}
-		if err != nil {
-			a.writeRootOrInternalError(w, recordingsHTTPOperationHistoricalRead, err)
-			return
-		}
-		start, err := historicalEventStart(result.Events, input.Params, input.StreamGenerationID)
-		if err != nil {
-			a.writeRootOrInternalError(w, recordingsHTTPOperationEventSubscribe, err)
-			return
-		}
-		streamGenerationID := input.StreamGenerationID
-		if strings.TrimSpace(streamGenerationID) == "" && len(result.Events) > 0 {
-			streamGenerationID = result.Events[0].Cursor.StreamGenerationID
-		}
-		a.streamFactoryEvents(
-			w, r, historicalEventSubscription(result.Events[start:]),
-			input.SessionID, streamGenerationID, len(result.Events[start:]),
-		)
-		return
+		a.probeHistoricalEventStreamRecovery(w, r, input)
+		return true
 	}
+	a.probeEventStreamRecovery(w, input)
+	return true
+}
+
+func (a *Adapter) handleHistoricalEvents(
+	w http.ResponseWriter,
+	r *http.Request,
+	input EventSubscribeInput,
+) bool {
+	if !isDurableHistorySession(input.SessionID) {
+		return false
+	}
+	result, err := a.historicalRecording(r.Context(), input.SessionID)
+	if shouldEndOnRequestContext(r.Context(), err) {
+		return true
+	}
+	if err != nil {
+		a.writeRootOrInternalError(w, recordingsHTTPOperationHistoricalRead, err)
+		return true
+	}
+	start, err := historicalEventStart(result.Events, input.Params, input.StreamGenerationID)
+	if err != nil {
+		a.writeRootOrInternalError(w, recordingsHTTPOperationEventSubscribe, err)
+		return true
+	}
+	streamGenerationID := input.StreamGenerationID
+	if strings.TrimSpace(streamGenerationID) == "" && len(result.Events) > 0 {
+		streamGenerationID = result.Events[0].Cursor.StreamGenerationID
+	}
+	a.streamFactoryEvents(
+		w, r, historicalEventSubscription(result.Events[start:]),
+		input.SessionID, streamGenerationID, len(result.Events[start:]),
+	)
+	return true
+}
+
+func (a *Adapter) handleLiveEvents(
+	w http.ResponseWriter,
+	r *http.Request,
+	input EventSubscribeInput,
+) {
 	request, err := SubscribeRequestFromAPI(input)
 	if err != nil {
 		a.writeError(w, http.StatusBadRequest, "invalid event reconnect cursor", "BAD_REQUEST")
